+++ src/components/Layout/Layout.css
@@ -26,6 +26,7 @@ locator-layout {
 
   &::part(main) {
     grid-area: main;
+    overflow: auto;
   }
 
   &::part(aside) {
