+++ src/styles/tokens/button.css
@@ -1,6 +1,6 @@
 :host {
   --diamond-button-background: var(--color-white);
-  --diamond-button-background-hover: var(--diamond-button-background);
+  --diamond-button-background-hover: var(--theme-background-accent);
   --diamond-button-background-disabled: var(--diamond-button-background);
   --diamond-button-border-color: var(--color-green);
   --diamond-button-border-color-hover: var(--color-green-dark);
@@ -8,6 +8,7 @@
   --diamond-button-color: var(--diamond-button-border-color);
   --diamond-button-color-hover: var(--diamond-button-border-color-hover);
   --diamond-button-color-disabled: var(--diamond-button-border-color-disabled);
+  --diamond-button-padding-block: 0.75rem;
 
   --diamond-button-primary-background: var(--color-green);
   --diamond-button-primary-background-hover: var(--color-green-dark);
@@ -15,4 +16,6 @@
   --diamond-button-primary-color: var(--color-white);
   --diamond-button-primary-color-hover: var(--diamond-button-primary-color);
   --diamond-button-primary-color-disabled: var(--diamond-button-primary-color);
+
+  --diamond-button-text-background-hover: var(--theme-background-accent);
 }
